feat: wild stab at embarassing problem
I've been having trouble with a weird bug where,
no matter how I configure the kudzu.news domain
that I bought, I can't make it actually work. It
takes you to the app just fine, _but_ it keeps
redirecting back to the Heroku domain.

I haven't googled this or anything, but I was using
the bathroom just now and realized that the problem
might be caused by this setting. So here's a stab
in that direction. #yolo